Split vLLM benchmark suite
